Update README Add beeps Tune debounce and beep intervals

+++ src/Bluenumi/Melody.cpp
@@ -17,6 +17,7 @@
 
 #include "Melody.h"
 
-int TONE_UP_NOTES[] = {NOTE_A4};
-int TONE_UP_DURATIONS[] = {DUR_QT};
+uint16_t TONE_UP_NOTES[] = {NOTE_A4};
+uint16_t TONE_UP_DURATIONS[] = {DUR_QT};
 Melody TONE_UP_MELODY = Melody(TONE_UP_NOTES, TONE_UP_DURATIONS);
+
